feat($env): 添加babel
