+++ package.json
@@ -2,9 +2,15 @@
   "name": "price-transfer",
   "version": "1.0.0",
   "description": "将数字转换以价格的形式进行展示",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "./node_modules/.bin/mocha --require intelli-espower-loader"
+    "test": "./node_modules/.bin/mocha --require intelli-espower-loader",
+    "compile": "babel --presets es2015 -d lib/ src/",
+    "prepublish": "npm run compile"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-preset-es2015": "^6.24.1"
   },
   "dependencies": {
     "intelli-espower-loader": "^1.0.1",
